improve visual appearance, add help message

+++ rr.sh
@@ -6,9 +6,12 @@ BOLD="\e[1m"
 ERROR="\e[1;31m"
 WARNING="\e[1;38;5;208m"
 SUCCESS="\e[1;32m"
+EM="\e[1;34m"
 INFO="\e[1;35m"
 RESET="\e[0m"
 
+terminal=$(tty)
+columns=$(stty -a <"$terminal" | egrep -Eo '\d+ columns;' | egrep -Eo '\d+')
 
 h1() {
     hx "━" "$1"
@@ -31,10 +34,10 @@ log() {
 }
 
 trace() {
-    printf "━%.0s" {1..80}
+    printf "━%.0s" {1..$columns}
     printf "\n"
     echo "$*"
-    printf "─%.0s" {1..80}
+    printf "─%.0s" {1..$columns}
     printf "\n"
     $1 "${@:2}"
     echo
@@ -62,8 +65,12 @@ check() {
     trace cargo clippy
     trace cargo clippy --all-features
     trace cargo test
-    trace npm run pyright
-    trace run-python-tests
+    if [ -d prettypretty ]; then
+        trace npm run pyright
+    fi
+    if [ -d test ]; then
+        trace run-python-tests
+    fi
 }
 
 build() {
@@ -72,13 +79,42 @@ build() {
 }
 
 docs() {
-    trace mdbook build docs
+    if [ -d docs ]; then
+        trace mdbook build docs
+    fi
+
     trace cargo rustdoc --all-features -- -e $(realpath docs/pretty.css)
-    trace sphinx-build -a -b html docs target/doc/python
-    trace rm -rf target/doc/python/.doctrees
+
+    if [ -d docs ]; then
+        trace sphinx-build -a -b html docs target/doc/python
+        trace rm -rf target/doc/python/.doctrees
+    fi
 }
 
-case $1 in
+help() {
+    echo "${EM}./rr.sh [check|build|docs|all|help]${RESET}"
+    echo
+    echo "${BOLD}check${RESET} :  Check that the source code is well-formatted,"
+    echo "         free of lint, and altogether in good shape."
+    echo "${BOLD}build${RESET} :  Build and locally install the Python extenion module."
+    echo "${BOLD}docs${RESET}  :  Build the user guide, the Rust API documentation,"
+    echo "         and the Python API documentation; see target/doc"
+    echo "${BOLD}all${RESET}   :  Perform all of the above tasks in the listed order."
+    echo "${BOLD}help${RESET}  :  Show this help message and exit."
+    exit 1
+}
+
+target="$1"
+
+if [ -z "$1" ]; then
+    target="build"
+fi
+
+
+case $target in
+    "-h" | "--help" )
+        help
+        ;;
     check )
         check
         ;;
@@ -94,7 +130,7 @@ case $1 in
         docs
         ;;
     * )
-        log ERROR "\"$1\" is not a valid run target!"
+        log ERROR "\"$target\" is not a valid runner target!"
         exit 1
         ;;
 esac
